requirements.txt
langchain==0.1.20
openai==1.29.0
streamlit==1.34.0
pinecone-client==4.1.0